Amend secret_key

+++ run_local.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+
+if [ -z ${SECRET_KEY} ]; then
+    echo "SECRET_KEY is unset."
+    exit 1
+fi
+
 docker compose down --remove-orphans
 docker compose up -d --build
 docker compose wait migrations
